requirements.txt
apache-airflow-providers-mongo==5.0.1
apache-airflow-providers-redis==4.0.1
Telethon==1.39.0
transformers==4.49.0
pika==1.3.2
pendulum==3.0.0
feedparser==6.0.11
web3==7.9.0
kafka-python==2.1.2
hexbytes==1.3.0
pymongo==4.11.3
torch==2.6.0
hf-xet==1.0.2
pandas==2.2.3